LPS-95166 Move UserGroupGroupRoleFinderTest of the com.liferay.portal.service.persistence package to module
